pyproject.toml
[tool.poetry]
name = "iot-zeroconf-service"
version = "0.1.0"
description = ""
authors = ["Bhaktij Koli <bhaktijkoli121@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.6.0"
python-dotenv = "^0.21.0"
zeroconf = "^0.39.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
